[MIG] currency_rate_update: Migration to 8.0
Move directory to main folder.

Make the module instalable.

Update company to new api.

Update res currency update services to new api.

Move init file to model.

Update copyrights, colaborators will be visible to readme file.

Add readme.md, cron data for services, modify declaration of views.

Modify views.

Added abstract classes, split services in different files.

Remove self.ensure_one from @api.one.

Modify contribuitors.

Update acces rights.

Update openerp.py security file.

Re-add imports of services abstract classes.

Add dynamic import of services to pass flake8.

Remove trailing whitespaces.

Update datetime.time.min method not existent to datetime.min.time().

Calculate midnight based on time.

Update travis relative imports, warnings.

Update code according to comments.

Update CA_BoC messages.
