+++ currency_rate_update/i18n/ab.po
@@ -1,0 +1,180 @@
+# Translation of OpenERP Server.
+# This file contains the translation of the following modules:
+#	* c2c_currency_rate_update
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: OpenERP Server 5.0.0\n"
+"Report-Msgid-Bugs-To: support@openerp.com\n"
+"POT-Creation-Date: 2013-10-18 17:52+0000\n"
+"PO-Revision-Date: 2013-11-11 16:24+0000\n"
+"Last-Translator: Grzegorz Grzelak (Cirrus) <Unknown>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"X-Launchpad-Export-Date: 2014-06-12 06:31+0000\n"
+"X-Generator: Launchpad (build 17041)\n"
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Currency auto update configuration"
+msgstr "Aktualizacja kursów walut"
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Currency updates services"
+msgstr "Serwisy aktualizacji kursów"
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,company_id:0
+msgid "linked company"
+msgstr "dla firmy"
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Rate"
+msgstr "Kurs"
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Rates"
+msgstr "Kursy"
+
+#. module: currency_rate_update
+#: field:res.company,multi_company_currency_enable:0
+msgid "Multi company currency"
+msgstr "Waluta dla wielu firm"
+
+#. module: currency_rate_update
+#: field:res.company,interval_type:0
+msgid "Currency update frequence"
+msgstr "Częstotliwość aktualizacji kursów"
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Day(s)"
+msgstr "Codziennie"
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,currency_to_update:0
+msgid "currency to update with this service"
+msgstr "Waluty aktualizowane tym serwisem"
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Narodowy Bank Polski"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,note:0
+msgid "update notice"
+msgstr "Notatka aktualizacji"
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,service:0
+msgid "Webservice to use"
+msgstr "Strona web do zastosowania"
+
+#. module: currency_rate_update
+#: help:currency.rate.update.service,max_delta_days:0
+msgid ""
+"If the time delta between the rate date given by the webservice and the "
+"current date exeeds this value, then the currency rate is not updated in "
+"OpenERP."
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,services_to_use:0
+msgid "Currency update services"
+msgstr "Serwisy aktualizacji kursów"
+
+#. module: currency_rate_update
+#: constraint:currency.rate.update.service:0
+msgid "'Max delta days' must be >= 0"
+msgstr ""
+
+#. module: currency_rate_update
+#: code:addons/currency_rate_update/currency_rate_update.py:89
+#: sql_constraint:currency.rate.update.service:0
+#, python-format
+msgid "You can use a service one time per company !"
+msgstr "Możesz stosować jeden serwis na firmę !"
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Yahoo Finance "
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,max_delta_days:0
+msgid "Max delta days"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Admin.ch"
+msgstr ""
+
+#. module: currency_rate_update
+#: model:ir.model,name:currency_rate_update.model_res_company
+msgid "Companies"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Currencies to update with this service"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Week(s)"
+msgstr "Co tydzień"
+
+#. module: currency_rate_update
+#: help:res.company,multi_company_currency_enable:0
+msgid ""
+"if this case is not check you can not set currency is active on two company"
+msgstr ""
+"Jeśli nie zaznaczysz, to nie będziesz mógł ustawić waluty jako aktywnej dla "
+"dwóch firm"
+
+#. module: currency_rate_update
+#: code:addons/currency_rate_update/currency_rate_update.py:153
+#: model:ir.model,name:currency_rate_update.model_currency_rate_update
+#: model:ir.model,name:currency_rate_update.model_currency_rate_update_service
+#, python-format
+msgid "Currency Rate Update"
+msgstr "Aktualizacja kursów walut"
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Logs"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,auto_currency_up:0
+msgid "Automatical update of the currency this company"
+msgstr "Automatyczna aktualizacja kursów walut dla tej firmy"
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Month(s)"
+msgstr "Co miesiąc"
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "European Central Bank"
+msgstr ""
+
+#. module: currency_rate_update
+#: help:res.company,interval_type:0
+msgid ""
+"changing this value will\n"
+"                                                 also affect other compagnies"
+msgstr "Zmiana tej wartości wpłynie również na inne firmy"
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Refresh currencies"
+msgstr "Aktualizuj kursy"
