+++ currency_rate_update/i18n/pt_BR.po
@@ -1,0 +1,179 @@
+# Brazilian Portuguese translation for account-financial-tools
+# Copyright (c) 2014 Rosetta Contributors and Canonical Ltd 2014
+# This file is distributed under the same license as the account-financial-tools package.
+# FIRST AUTHOR <EMAIL@ADDRESS>, 2014.
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: account-financial-tools\n"
+"Report-Msgid-Bugs-To: FULL NAME <EMAIL@ADDRESS>\n"
+"POT-Creation-Date: 2013-10-18 17:52+0000\n"
+"PO-Revision-Date: 2014-01-10 11:21+0000\n"
+"Last-Translator: FULL NAME <EMAIL@ADDRESS>\n"
+"Language-Team: Brazilian Portuguese <pt_BR@li.org>\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"X-Launchpad-Export-Date: 2014-06-12 06:31+0000\n"
+"X-Generator: Launchpad (build 17041)\n"
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Currency auto update configuration"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Currency updates services"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,company_id:0
+msgid "linked company"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Rate"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Rates"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,multi_company_currency_enable:0
+msgid "Multi company currency"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,interval_type:0
+msgid "Currency update frequence"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Day(s)"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,currency_to_update:0
+msgid "currency to update with this service"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Narodowy Bank Polski"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,note:0
+msgid "update notice"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,service:0
+msgid "Webservice to use"
+msgstr ""
+
+#. module: currency_rate_update
+#: help:currency.rate.update.service,max_delta_days:0
+msgid ""
+"If the time delta between the rate date given by the webservice and the "
+"current date exeeds this value, then the currency rate is not updated in "
+"OpenERP."
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,services_to_use:0
+msgid "Currency update services"
+msgstr ""
+
+#. module: currency_rate_update
+#: constraint:currency.rate.update.service:0
+msgid "'Max delta days' must be >= 0"
+msgstr ""
+
+#. module: currency_rate_update
+#: code:addons/currency_rate_update/currency_rate_update.py:89
+#: sql_constraint:currency.rate.update.service:0
+#, python-format
+msgid "You can use a service one time per company !"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Yahoo Finance "
+msgstr ""
+
+#. module: currency_rate_update
+#: field:currency.rate.update.service,max_delta_days:0
+msgid "Max delta days"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "Admin.ch"
+msgstr ""
+
+#. module: currency_rate_update
+#: model:ir.model,name:currency_rate_update.model_res_company
+msgid "Companies"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Currencies to update with this service"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Week(s)"
+msgstr ""
+
+#. module: currency_rate_update
+#: help:res.company,multi_company_currency_enable:0
+msgid ""
+"if this case is not check you can not set currency is active on two company"
+msgstr ""
+
+#. module: currency_rate_update
+#: code:addons/currency_rate_update/currency_rate_update.py:153
+#: model:ir.model,name:currency_rate_update.model_currency_rate_update
+#: model:ir.model,name:currency_rate_update.model_currency_rate_update_service
+#, python-format
+msgid "Currency Rate Update"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:currency.rate.update.service:0
+msgid "Logs"
+msgstr ""
+
+#. module: currency_rate_update
+#: field:res.company,auto_currency_up:0
+msgid "Automatical update of the currency this company"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:res.company,interval_type:0
+msgid "Month(s)"
+msgstr ""
+
+#. module: currency_rate_update
+#: selection:currency.rate.update.service,service:0
+msgid "European Central Bank"
+msgstr ""
+
+#. module: currency_rate_update
+#: help:res.company,interval_type:0
+msgid ""
+"changing this value will\n"
+"                                                 also affect other compagnies"
+msgstr ""
+
+#. module: currency_rate_update
+#: view:res.company:0
+msgid "Refresh currencies"
+msgstr ""
