+++ currency_rate_update/security/ir.model.access.csv
@@ -1,0 +1,3 @@
+id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
+access_account_currency_rate_update_service_settings,currency.rate.update.service,model_currency_rate_update_service,account.group_account_manager,1,1,1,1
+access_account_currency_rate_update_service_financial,currency.rate.update.service,model_currency_rate_update_service,base.group_system,1,1,1,1
