+++ currency_rate_update/services/__init__.py
@@ -18,5 +18,3 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-
-from . import model
