+++ currency_rate_update/services/update_service_MX_BdM.py
@@ -1,0 +1,82 @@
+# -*- coding: utf-8 -*-
+##############################################################################
+#
+#    Copyright (c) 2009 CamptoCamp. All rights reserved.
+#    @author Nicolas Bessi
+#
+#    Abstract class to fetch rates from Banco de México
+#
+#    This program is free software: you can redistribute it and/or modify
+#    it under the terms of the GNU Affero General Public License as
+#    published by the Free Software Foundation, either version 3 of the
+#    License, or (at your option) any later version.
+#
+#    This program is distributed in the hope that it will be useful,
+#    but WITHOUT ANY WARRANTY; without even the implied warranty of
+#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#    GNU Affero General Public License for more details.
+#
+#    You should have received a copy of the GNU Affero General Public License
+#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+#
+##############################################################################
+
+from .currency_getter_interface import Currency_getter_interface
+
+import logging
+_logger = logging.getLogger(__name__)
+
+
+class MX_BdM_getter(Currency_getter_interface):
+    """Implementation of Currency_getter_factory interface
+    for Banco de México service
+
+    """
+
+    def rate_retrieve(self):
+        """ Get currency exchange from Banxico.xml and proccess it
+        TODO: Get correct data from xml instead of process string
+        """
+        url = ('http://www.banxico.org.mx/rsscb/rss?'
+               'BMXC_canal=pagos&BMXC_idioma=es')
+
+        from xml.dom.minidom import parse
+        from StringIO import StringIO
+
+        logger = logging.getLogger(__name__)
+        logger.debug("Banxico currency rate service : connecting...")
+        rawfile = self.get_url(url)
+
+        dom = parse(StringIO(rawfile))
+        logger.debug("Banxico sent a valid XML file")
+
+        value = dom.getElementsByTagName('cb:value')[0]
+        rate = value.firstChild.nodeValue
+
+        return float(rate)
+
+    def get_updated_currency(self, currency_array, main_currency,
+                             max_delta_days=1):
+        """implementation of abstract method of Curreny_getter_interface"""
+        logger = logging.getLogger(__name__)
+        # we do not want to update the main currency
+        if main_currency in currency_array:
+            currency_array.remove(main_currency)
+
+        # Suported currencies
+        suported = ['MXN', 'USD']
+        for curr in currency_array:
+            if curr in suported:
+                # Get currency data
+                main_rate = self.rate_retrieve()
+                if main_currency == 'MXN':
+                    rate = 1 / main_rate
+                else:
+                    rate = main_rate
+            else:
+                # No other currency supported
+                continue
+
+            self.updated_currency[curr] = rate
+            logger.debug("Rate retrieved : %s = %s %s" %
+                         (main_currency, rate, curr))
