+++ currency_rate_update/view/currency_rate_update.xml
@@ -1,0 +1,44 @@
+<?xml version="1.0"?>
+<openerp>
+    <data>
+        <record model="ir.ui.view" id="currency_rate_update_tree">
+            <field name="name">Update Rates service</field>
+            <field name="model">currency.rate.update.service</field>
+            <field name="arch" type="xml">
+                <tree string="Rates">
+                    <field name="service"/>
+                    <field name="currency_to_update"/>
+                    <field name="interval_type"/>
+                    <field name="interval_number"/>
+                    <field name="next_run"/>
+                </tree>
+            </field>
+        </record>
+
+        <record model="ir.ui.view" id="currency_rate_update_form">
+            <field name="name">Update Rates</field>
+            <field name="model">currency.rate.update.service</field>
+            <field name="arch" type="xml">
+                <form string="Rate">
+                    <group>
+						<group>
+                            <field name="service"/>
+                            <field name="max_delta_days"/>
+                        </group>
+                        <group>                            
+                            <field name="interval_type"/>
+                            <field name="interval_number"/>
+                            <field name="next_run"/>
+                        </group>
+                    </group>
+                    <separator string="Currencies to update with this service" colspan="4"/>
+                    <field name="currency_list" invisible="1"/>
+                    <field name="currency_to_update" colspan="4" nolabel="1" domain="[('id','in', currency_list[0][2])]"/>
+                    <separator string="Logs" colspan="4"/>
+                    <field name="note" colspan="4" nolabel="1"/>
+                </form>
+            </field>
+        </record>
+
+    </data>
+</openerp>
